daemon/container,compartment: allow to join net namespaces of a parent
We introduce the concept of a parent container/compartment. It is now
possible to join network namespaces of such a parent. The corresponding
name of a network namespace which was built inside of the parent, e.g.,
by 'ip netns add <name>' needs to be set as configuration parameter
'parent_netns' in the child's container config.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>