daemon/c_seccomp: Introduce ioctl() emulation for RTC handling
We implement a partial filter which continues all system calls
and only handles (notify) ioctl() calls related to the RTC.

Remember that this needs to be integrated in the overall filter
since we assume syscall nr already loaded and if it is not
SYS_ioctl, we jump one instruction beyond the last instruction
of filter_ioctl to the next comparison of syscall number.

We identify the correct ioctl by the 'cmd' parameter.
This could collide with other ioctl() calls unrelated to RTC
which define the same 'number' as their command.

Thus, we check in the emulation if the opened fd corresponds
to a real time clock device and if not just exit emulation
early and let the kernel syscall continue as is.

In any case of error, we just continue the syscall in the kernel.

The emulation, further checks valid CAP_SYS_TIME and cgroup
device access. We also join the corresponding namespaces
of the container for executing the ioctl() on behalf of the
container with minimal rights only retaining CAP_SYS_TIME
and switching to the uid of the container root user in the
initial user namespace.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>